requirements.txt
fastapi[all]
sqlalchemy
databases
asyncpg
alembic
pydantic
uvicorn
python-dotenv